requirements.txt
absl-py==1.0.0
aidungeonapi==0.0.6
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
discord.py==1.7.3
filelock==3.6.0
fire==0.4.0
fsspec==2022.2.0
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
gql==3.1.0
graphql-core==3.2.0
grpcio==1.44.0
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
Markdown==3.3.6
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.7
python-dotenv==0.19.2
pytorch-lightning==1.5.10
PyYAML==6.0
qrcode==7.3.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tokenizers==0.11.6
torch==1.11.0
torchmetrics==0.7.3
tqdm==4.63.1
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
yarl==1.7.2
